+++ packages/common-cosmos/src/codegen/util.ts
@@ -1,6 +1,7 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import path from 'path';
 import {loadFromJsonOrYaml} from '@subql/common';
 import {IDLObject} from 'wasm-ast-types';
 import {parseCosmosProjectManifest, ProjectManifestImpls} from '../project';
@@ -15,4 +16,8 @@ export function validateCosmosManifest(manifest: unknown): manifest is ProjectMa
 
 export function loadCosmwasmAbis(filePath: string): IDLObject {
   return loadFromJsonOrYaml(filePath) as IDLObject;
+}
+
+export function tmpProtoDir(tmpDir: string, protoPath: string): string {
+  return path.join(tmpDir, path.basename(protoPath));
 }