+++ packages/common-cosmos/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Update codegen pathing for OS consistency (#196)
 
 ## [3.2.0] - 2023-11-01
 ### Added